Added a few comments
