--- pages/drugs.tsx
@@ -1,81 +1,0 @@
-interface drugList {
-  name: string;
-  packSize: string;
-  category: string;
-  price: number;
-}
-
-export const drugList = [
-  {
-    name: "Amlodipine 5mg",
-    packSize: "28",
-    category: "M",
-    price: "40",
-  },
-  {
-    name: "Amlodipine 10mg",
-    packSize: "28",
-    category: "M",
-    price: "80",
-  },
-  {
-    name: "Bisoprolol 1.25mg",
-    packSize: "28",
-    category: "M",
-    price: "20",
-  },
-  {
-    name: "Bisoprolol 2.5mg",
-    packSize: "28",
-    category: "M",
-    price: "40",
-  },
-  {
-    name: "Candesartan 4mg",
-    packSize: "28",
-    category: "M",
-    price: "180",
-  },
-  {
-    name: "Candesartan 8mg",
-    packSize: "28",
-    category: "M",
-    price: "360",
-  },
-  {
-    name: "Amlodipine 5mg",
-    packSize: "28",
-    category: "M",
-    price: "40",
-  },
-  {
-    name: "Amlodipine 10mg",
-    packSize: "28",
-    category: "M",
-    price: "80",
-  },
-  {
-    name: "Bisoprolol 1.25mg",
-    packSize: "28",
-    category: "M",
-    price: "20",
-  },
-  {
-    name: "Bisoprolol 2.5mg",
-    packSize: "28",
-    category: "M",
-    price: "40",
-  },
-  {
-    name: "Candesartan 4mg",
-    packSize: "28",
-    category: "M",
-    price: "180",
-  },
-  {
-    name: "Candesartan 8mg",
-    packSize: "28",
-    category: "M",
-    price: "360",
-  },
-];
